Use ANSI clearing codes instead of \033c
Mimic the behavior of using \033c by using \033[2J\033H instead.

- \033[2J: Clears the screen and (depending on implementation, I think?) returns the cursor to the home position.
- \033[H: Returns the cursor to the home position.

I haven't found reference to \033c in the documentation that I've perused, so I don't think that one was as standard as this one, either.

The combination of these two adequately resolves #1.  For me, no longer does the observed "phantom parrot" get rendered in the terminal.  It has the downside of sending a few extra bytes down the wire, but I don't think that's critically bad in this case. :grin:

[GitHub] This commit resolves #1.