Add all the parrots.
In this commit, I removed the original frames in the frames folder.
Then I created a script: make-parrots.sh
This script goes to https://cultofthepartyparrot.com/ and proceeds to
download all parrot gifs.

Then we split up the gifs using ImageMagick,
then we use image2ascii to convert each png to a txt file.

We scale all images to be 50 x 25.  That seems pretty close to the
original parrot.live size.

I modified index.js to read in all the frames
(I think performace wise, this is ok, since it is just bytes)

Then I modified the http server to allow urls with
parrot.live/<parrot>

In the url, if you don't specify a parrot, we show
the original parrot.

You only need to specify a parrot given the name, without
the word "parrot" in it.

We also now provide a /list endpoint.
This endpoint will show all the available parrot types.

Finally, I removed the color randomization for now
since all the parrots have color already embedded in the frames files.